@@ -43,7 +43,7 @@ if (typeof window !== 'undefined' && window.Vue) {
   install(window.Vue);
 };
 
-module.exports = {
+export default {
   install,
   version: '0.0.1'
 };
